ENH: make gcodeviewer to support G2 G3 arc move
For G2 and G3 command, use interpolation to generate
points between the start point and end point of arc.

Then, use the same way of drawing G1 extrusion to draw
arc.

Arc should be one drawing when user move the slider of
gcodeviewer. This also be supported in this patch.

Signed-off-by: salt.wei <salt.wei@bambulab.com>
Change-Id: I2219767b938797cb86b16dca16e47715b624729c