FIX: fix the empty layer in gcode for some model
Lazy raise was add to improve print quality in which
lift z was moved from change_layer() function to travel_to()
function.

But the first point of new layer may be same with last position
in 2D which cause travel_to() can't be hit and no lift z be
done when changing layer.

Add one flag to fix this issue.

Signed-off-by: salt.wei <salt.wei@bambulab.com>
Change-Id: I75f461437c16e0942c673d418ba4f31c7816f6a1