ENH: add slope for long slanting travel with z lift
sloping travel was used to replace "lift and move" to reduce
stringing. But if the slope is too flat, the z hight is close
to the printed surface. Than measn the z lift is almost useless
to avoid hitting some warping place.

Use a "slope and move" travel instead to make z lift more rapid
than slope only.

Signed-off-by: salt.wei <salt.wei@bambulab.com>
Change-Id: I0a0d1741c14231aa6b2191292d0449369850d296