ENH: remove the anchor at the start of sparse infill
The infill area print sequence is not fixed, which means that
sparse infill may be printed after solid infill. So there has
posibility that the start anchor of sparse infill is 100% overhang
and not supported.

Starting printing at air is unsafe and has bed influence on print
quality. Remove the useless start anchor which other slicer doesn't
have as well.

Signed-off-by: salt.wei <salt.wei@bambulab.com>
Change-Id: I0a12284f1ae57b7ba25ff99ffd1450134f935c12