ENH: disable slow down for overhang when fuzzy skin is used
When fuzzy skin is used, it's unmeaning to calculate
the overhang degree of extrusion line, because there
are a lot of random fuzzy point which change the line.

Disable in this case. This can also speed slicing up
when fuzzy skin is used.

Signed-off-by: salt.wei <salt.wei@bambulab.com>
Change-Id: Ia65e20e24fd7a5cd224635fd69a272387135ba12
(cherry picked from commit 93ab1b521ed238a138412f1efe4987cd56ac6858)