Bambu GUI: be much more careful about calling Layout() and Refresh()
Repainting the screen is expensive and causes lots of idle CPU usage when
the printer is continuously sending updates over the network (or a timer is
simply firing); in some cases, it can even cause a SVG to continuously be
rerendered to a bitmap.  In general, we should avoid causing widgets to
repaint if they don't change; we add some checks to see if anything has
actually changed inside of a widget before we actually commit to repainting.

This is a first step towards taking much finer-grained control over when
these operations are called, rather than unconditionally calling them at
1Hz.  In the future, we should remove the 1Hz timer entirely -- and only
call Layout on things that need their layout updated -- but in the mean
time, this change substantially reduces the amount of CPU time (and
battery!) that Bambu Studio consumes in the background when nothing is
happening (particularly important for laptop users!).