uncomment 'filament used' lines in gcode generator
I'm looking to build third-party tools that are compatible with Bambu Studio. One of the metrics that I'd like to track is the amount of filament used. Both PrusaSlicer and OrcaSlicer emit this in the exact same way, but Bambu Studio comments it out; it'd be nice to revert this.