unix: Fix "Open in BambuStudio" button on MakerWorld
Make the "Open in BambuStudio" button on MakerWorld works by informing
the system that BambuStudio is the application to handle URLs with the
bambustudio scheme, and that it can handle URLs, not just local files.

This fixes this error message visible in Firefox' console when clicking
the button on the website:
Prevented navigation to “bambustudio://open?file=https%3A%2F%2Fmakerworld.bblmw.com[...]bambu_scraper_grip_ams.3mf” due to an unknown protocol.