ENH: linux.d/debian: use correct string match syntax
The test(1) command (also known as '[') does not use
the '==' syntax for testing equality, which is a
feature of bash's '[[' builtin. String equality
is tested with '=', and numeric equality with '-eq'.
This eliminates a series of warnings when running
BuildLinux.sh on Debian-derived systems:

[schwarzgerat](0) $ ./BuildLinux.sh -u
./linux.d/debian: line 39: [: ==: unary operator expected
./linux.d/debian: line 39: [: ==: unary operator expected
./linux.d/debian: line 42: [: ==: unary operator expected

github pull request: https://github.com/bambulab/BambuStudio/pull/5103

Change-Id: I44f41f29825b139dfc0125825a275aee4fa4e739