Added Linux Mint support to the build system.
Created symlinks for `linuxmint` and `ubuntu` in `linux.d`, and removed
the Ubuntu check in `BuildLinux.sh`.  This is an easy way to handle
derivative distributions.

One particular caveat: symlinks don't play well
with older versions of git on Windows.  It's unlikely to be a problem
unless somebody tries to commit changes to `linux.d` from windows.
See: https://stackoverflow.com/questions/5917249/git-symbolic-links-in-windows

An alternative approach is to use the `ID_LIKE` tag in
`/etc/os-release`.  See example 3 in the `os-release` manual.