ENH: Change Linux build script to recognize more supported distributions
github pull request: https://github.com/bambulab/BambuStudio/pull/5686
This changes the way the Linux build script recognizes which
distribution script might be compatible with the host OS.  It
does so by checking both the `ID` and `ID_LIKE` tags in
`/etc/os-release` against available scripts in `linux.d`.  First match
is used.

Note that ID_LIKE isn't a 100% compatability guarantee, but it gives the
user a fighting chance that the script will work for their distribution,
and should reduce the need for manual addition of more distribution
targets.

Change-Id: Iadc6cdd328c3400ae48e7833fc5f924eb267e2a0