FIX: Fix focus stealing from app when filament changes
At every filament related event, the old code would try to set the
keyboard focus to the "Device" tab, leading to a number of, either,
focus stealing events, or notifications if the desktop has focus
stealing prevention (like GNOME).

Disable the focus setting on Linux to prevent that focus stealing.

Closes: https://github.com/bambulab/BambuStudio/issues/3046