Fix patch step for OCCT and OpenCV dependencies
The `PATCH_COMMAND` for OCCT and OpenCV are trying to patch files that
don't exist because of the addition of the `--directory` flag.
`PATCH_COMMAND` runs in the working directory of the downloaded source
files per my own testing. Specifying `--directory` here is causing the
patch to try to apply the patch to `dep_OpenCV-prefix/src/OpenCV` from
root of the downloaded source. `git apply` returns nonzero error codes
when patching non-existent files AND specifying the `--directory`
flag.