Fix Spanish translation for "Ludicrous"
The spanish .po file translates "Ludicrous" as "Lúdico", which
is absolutely erroneous. The correct translation is "Absurda",
as it was done when Spaceballs was dubbed both in Spain and
Latin American ("Ludicrous speed" -> "Velocidad absurda").

Of course, since, in Spanish, words have gender, all the other
speeds must be swapped from masculine to feminine to keep
the gender match.