FIX: Fix empty Prepare/Preview content in Flatpak
By removing the OSMesa dependency.

The "offscreen" rendering code was supposed to use OSMesa's offscreen
buffers so as to avoid an X11 dependency, but the CLI::run() function
called XInitThreads() every time, so there's very little chance that
this ever worked.

Remove the dependency so that we don't depend on functionality that's
been removed from newer mesa:
https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/33836

Closes: https://github.com/bambulab/BambuStudio/issues/7017
Closes: https://github.com/bambulab/BambuStudio/issues/6951