NEW: Allow priting Wall and Infill with different materials
The users has a requirement to use one material for filling and another material for printing the exterior wall (from the perspective of cost (reducing the cost of consumables for large filling) and playability (using softer material for filling on the exterior)).

We would like to thank orca for their inspiration on this feature, and we have referenced part of their implementation.

jira: STUDIO-14837
Change-Id: Iae102f2284152fe3c21726092a54ea5c69ebdb8e