FIX: This is a fix for a major macOS issue (Mouse input lost #7250)
On H2D/H2C, the filament-saver popup (FilamentGroupPopup) causes a mouse capture issue that renders the app non-responsive. The issue is with the Slice Plate SideButton object.

Because on macOS it doesn't always receive clean pairs of mouseDown and mouseReleased, it can get into a state where the mouse capture is stuck on that button, which means the button now eats up all mouse events for the app.

The fix (which should be VERY safe for all platforms) is to ensure that whenver we receive a mouseReleased, we always check if the button currently has a mouse capture, and then release it if it does (without checking the pressedDown state). I considered limiting the change to just macOS, but I really don't think it is necessary.

github pull request: https://github.com/bambulab/BambuStudio/pull/9188

Change-Id: Iea6035b092550ce47c369906591a6f0224001a91